refactor(studio): move version badge to bottom right corner
- Remove version display from Update section in navbar
- Create new subtle VersionBadge component in bottom right
- Use smaller font size (11px) with opacity transition
- More unobtrusive design as per team feedback
