+++ studio/frontend/src/components/navbar.tsx
@@ -123,20 +123,6 @@ function CopyableCommand({
   );
 }
 
-interface HardwareInfo {
-  gpu: {
-    gpu_name: string | null;
-    vram_total_gb: number | null;
-    vram_free_gb: number | null;
-  };
-  versions: {
-    unsloth: string | null;
-    torch: string | null;
-    transformers: string | null;
-    cuda: string | null;
-  };
-}
-
 function UpdateStudioInstructions({
   className,
   defaultShell,
@@ -147,7 +133,6 @@ function UpdateStudioInstructions({
   showTitle?: boolean;
 }): ReactElement {
   const [shell, setShell] = useState<UpdateShell>(defaultShell);
-  const [version, setVersion] = useState<string | null>(null);
   const prefersReducedMotion = useReducedMotion();
   const windows = shell === "windows";
   const fadeTransition = prefersReducedMotion
@@ -161,16 +146,6 @@ function UpdateStudioInstructions({
     setShell(defaultShell);
   }, [defaultShell]);
 
-  useEffect(() => {
-    fetch("/api/system/hardware")
-      .then((response) => {
-        if (!response.ok) throw new Error("Failed to fetch");
-        return response.json();
-      })
-      .then((data: HardwareInfo) => setVersion(data.versions.unsloth))
-      .catch(() => {});
-  }, []);
-
   return (
     <div className={cn("flex flex-col gap-3", className)}>
       {showTitle ? (
@@ -178,14 +153,6 @@ function UpdateStudioInstructions({
           <p className="shrink-0 whitespace-nowrap text-sm font-semibold font-heading">
             Update Unsloth Studio
           </p>
-          <div className="flex items-center gap-1.5 text-sm">
-            <span className="font-mono text-muted-foreground">
-              {version ? `v${version}` : "..."}
-            </span>
-            <span className="font-extrabold tracking-[0.12em] text-primary">
-              BETA
-            </span>
-          </div>
         </div>
       ) : null}
       <div className="flex items-center justify-end gap-0.5 text-[11px]">
